Initial addition of plugins to click CLI interface.
Help works.
