+++ package_algorithms/algorithms_outer2.py
@@ -4,6 +4,9 @@ from algorithm_tester.tester_dataclasses import Task, Solution, Thing, ConfigCou
 class Greedy_Outer(Algorithm):
     """ Uses Brute force  """
 
+    def get_name(self) -> str:
+        return "Greedy_Outer"
+
     def get_column_descriptions(self, show_time: bool = True):
         res: str = super().get_column_descriptions(show_time=show_time)
         res[-2] = "|replacedGreedy|"
